Initialize I2C buses, attach FeRAM [rffe-board]
Initialize I2C0 and I2C1 buses. I2C0 is exposed through the character
driver interface to allow the rffe application interact directly with
the CDCE906 PLL (not the best solution).

The I2C1 bus is used to access the board's FeRAM. The FeRAM is mapped
under the '/dev/feram0' file.