SCPI over TCP implementation [rffe-app]
This is a huge commit, mainly due the inclusion of libscpi. The rffe
application now supports the SCPI protocol via TCP/IP. For now the TCP
port is hardcoded to 9001, but this can be changed in future versions.

Only commands for reading the temperature of channels AC and BD are
implemented now in addition to the mandatory SCPI std v1999.0 4.1.1
commands.