Firmware update support [rffe-board rffe-app]
The older firmware based on mbed os included a bootloader that will
copy firmware updates from the flash region starting from 0x48000 into
0x10000. This bootloader uses the first 64KiB from program memory, so
the linker script has to be modified to take this into consideration.

The firmware update server runs under the 9090 port and accepts the
following commands:

- e (character 'e', erases the flash from 0x48000 to 0x7FFFF);
- w (character 'w', Writes 256 bytes to the flash. Expects 4 bytes of
   the starting address in little endian format, and 256 bytes of the
   data to be written to flash);
- r (character 'r', resets the microcontroller);

If a command is executed successfully, firmware update server will
answer with '1' (character 1), except for the reset command (that will
reset the microcontroller immediately). Unrecognized commands will be
answered with a '0' (character 0).