Don't erase the bootloader when flashing with openocd
OpenOCD will erase the whole flash if the program command is invoked
with an ELF binary, even if it the elf file doesn't contain data in
the bootloader section (0x0000 - 0xFFFF). A workaround for this issue
is to use the raw binary file and instruct openocd to write it
on flash starting from the address 0x10000.