Set watchdog overflow flag before reset [rffe-board]
Issuing a system reset upon calling up_systemreset() would cause the
microcontroller to enter in bootloader mode if the DTR and RTS lines
are high (which happens often as these lines are floating dua a
hardware bug). Setting the WDTOF flag before reseting informs the
bootloader to jump directly to application instead of waiting for ISP
commands.