Fix and improve error message information [rffe-app]
sockfd was being passed to the printf calls in place of the thread ID.
While this version of NuttX does not yet have gettid(), its getpid()
actually implements the semantics of gettid(), and returns the thread
ID, so we can use it in the error messages.

Furthermore, the return value for recv() would always be -1, so printing
errno is what helps us in determining the error cause.