Try to workaround the socket exhaustion bug in the SCPI task [rffe-app]
Since the new RFFE IOC [1] has been deployed, we observed that after
some time some RFFEs would stop accepting new connections from the
IOC, even though the other connections were supposedly closed. We
suspected initially that the socket timeout on recv() stopped working
after a while, but after some debugging it seems that the problem
might be in the close() function, that blocks for an unusual amount of
time (more than a minute in some cases) if it doesn't receive the
socket close from the client side.

Opening 4 connections using rffe_nuttx_lib.py library and trying to
open a 5th connection with the IOC seems to reproduce the problem (one
of the 4 sockets never closes, even when waiting for a long
time). This suggests a potential concurrency bug in the Nuttx RTOS.

Only one SCPI connection is necessary in practice, so limiting the
SCPI server task to accepting only 1 connection might workaround this
bug, but further testing is needed.

[1]: https://github.com/lnls-dig/rffe-epics-ioc